winmm: compute the CD audio playback range better
The final track of a disc (e.g. Lo Wang's Rap) could not be played
because the end position was being set as the start of the following
track, which doesn't exist.

This is somewhat of a fools errand as for CD audio, apparently since
Vista, MCI_NOTIFY has not worked properly so we never know to restart
playback for looping music.